Implement user purchase validation and course access control
This commit adds the logic in page.tsx to validate if a user has made a purchase for the course they are trying to access. If not, the user is redirected to the checkout page for that course. If the course does not exist, a 'not-found' page is returned. This commit also moves the getUserIdFromEmail and getPurchasedCourses functions from auth-options.tsx to queries.ts. Furthermore, LearnPage() in page.tsx no longer depends on the session to map over the purchased courses. The course_id is also being selected now in getProductDetails.
