+++ src/lib/auth/auth-options.tsx
@@ -1,6 +1,7 @@
 import { sql } from "@vercel/postgres";
 import { type AuthOptions } from "next-auth"
 import GithubProvider from "next-auth/providers/github"
+import { getUserIdFromEmail, getPurchasedCourses } from "@/lib/queries";
 
 declare module "next-auth" {
   interface Profile {
@@ -78,16 +79,4 @@ export const authOptions = {
   }
 } as AuthOptions;
 
-async function getUserIdFromEmail(email: string): Promise<number | null> {
-  // Implement logic to fetch user ID from the database based on email
-  // Return the user ID or null if not found
-  const userRes = await sql`SELECT student_id FROM students WHERE email = ${email}`;
-  return userRes.rowCount > 0 ? userRes.rows[0].student_id : null;
-}
 
-async function getPurchasedCourses(userId: number): Promise<number[]> {
-  // Implement logic to fetch purchased courses from the database
-  // Return an array of course IDs or details
-  const purchasesRes = await sql`SELECT course_id from CourseEnrollments WHERE student_id = ${userId}`;
-  return purchasesRes.rows.map(row => row.course_id);
-}
